Move meta types to new file
There's some pretty complex type logic defining Remote and Local that is in
async_task.ts somewhat arbitrarily. This moves it to a new file called types.ts.
This makes it a little easier to understand the project layout.
